requirements.txt
torch~=2.0.1
numpy~=1.21.6
pillow~=10.0.0